+++ src/main/java/com/jsp/wms/responsedto/InventoryResponse.java
@@ -15,7 +15,7 @@ import com.jsp.wms.enums.MaterialTypes;
 @NoArgsConstructor
 @Builder
 public class InventoryResponse {
-    private Long inventoryId;
+    private Integer inventoryId;
     private String productTitle;
     private Double lengthInMeters;
     private Double breadthInMeters;
